+++ static/js/click_handlers.js
@@ -15,6 +15,7 @@ import * as compose from "./compose";
 import * as compose_actions from "./compose_actions";
 import * as compose_state from "./compose_state";
 import * as emoji_picker from "./emoji_picker";
+import * as giphy from "./giphy";
 import * as hash_util from "./hash_util";
 import * as hashchange from "./hashchange";
 import * as hotspots from "./hotspots";
@@ -673,6 +674,27 @@ export function initialize() {
     });
 
     // COMPOSE
+
+    $("#compose_giphy_logo").on("click", () => {
+        $("#compose_box_giphy_grid").popover("toggle");
+        let gifs_grid = giphy.renderGIPHYGrid($("#giphy_grid_in_popover .popover-content")[0]);
+        $("body").on(
+            "keyup",
+            "#giphy-search-query",
+            // Use debounce to create a 400ms interval between
+            // every search. This makes the UX of searching pleasant
+            // by allowing user to finish typing before search
+            // begins.
+            _.debounce(() => {
+                gifs_grid.remove();
+                gifs_grid = giphy.update_grid_with_search_term();
+            }, 400),
+        );
+        $(document).one("compose_canceled.zulip compose_finished.zulip", () => {
+            gifs_grid.remove();
+            $("#compose_box_giphy_grid").popover("hide");
+        });
+    });
 
     // NB: This just binds to current elements, and won't bind to elements
     // created after ready() is called.